Revert "Fix Chess Battle Royal lobby matchmaking buckets"
