+++ bot/server.js
@@ -64,11 +64,6 @@ import {
   normalizeOnlineGameType
 } from './config/onlineGamePolicy.js';
 import { createCheckersRealtimeStore } from './utils/checkersRealtimeState.js';
-import {
-  buildLobbyMatchKey,
-  isMatchMetaCompatible,
-  normalizeMatchMeta
-} from './utils/lobbyMatchmaking.js';
 import { applyAuthoritativeMove, SIDES } from './utils/checkersAuthoritativeEngine.js';
 
 validateEnv();
@@ -482,6 +477,91 @@ const dominoLobbyConnectGraceMs = Number(process.env.DOMINO_LOBBY_CONNECT_GRACE_
 const checkersMoveRateLimitMs =
   Number(process.env.CHECKERS_MOVE_RATE_LIMIT_MS) || 120;
 
+const MATCH_META_KEYS = ['mode', 'playType', 'variant', 'targetPoints', 'tableSize', 'ballSet', 'token'];
+
+function normalizeTableSizeMeta(value = '') {
+  const normalized = String(value || '').trim().toLowerCase();
+  if (!normalized) return '';
+  if (normalized === 'pro') return '9ft';
+  if (normalized.includes('9') && normalized.includes('ft')) return '9ft';
+  if (normalized.includes('8') && normalized.includes('ft')) return '8ft';
+  if (normalized.includes('tournament')) return '9ft';
+  return normalized;
+}
+
+function normalizeBallSetMeta(value = '') {
+  const normalized = String(value || '').trim().toLowerCase();
+  if (!normalized) return '';
+  if (normalized === 'us' || normalized === 'usa') return 'american';
+  if (normalized === 'english') return 'uk';
+  return normalized;
+}
+
+function normalizeVariantMeta(value = '') {
+  const normalized = String(value || '').trim().toLowerCase();
+  if (!normalized) return '';
+  const compact = normalized.replace(/[\s_-]+/g, '');
+  if (normalized === 'us' || normalized === 'usa') return 'american';
+  if (normalized === 'english') return 'uk';
+  if (compact === 'eightball' || compact === '8ball') return '8ball';
+  if (compact === 'nineball' || compact === '9ball') return '9ball';
+  return normalized;
+}
+
+function normalizeMatchMetaValue(key, value) {
+  const normalized = String(value || '').trim().toLowerCase();
+  if (!normalized) return '';
+  if (key === 'tableSize') return normalizeTableSizeMeta(normalized);
+  if (key === 'ballSet') return normalizeBallSetMeta(normalized);
+  if (key === 'variant') return normalizeVariantMeta(normalized);
+  return normalized;
+}
+
+function normalizeMatchMeta(rawMeta = {}) {
+  const normalized = {};
+  MATCH_META_KEYS.forEach((key) => {
+    const value = rawMeta[key];
+    if (typeof value === 'string' && value.trim()) {
+      const normalizedValue = normalizeMatchMetaValue(key, value);
+      if (normalizedValue) normalized[key] = normalizedValue;
+    }
+  });
+  return normalized;
+}
+
+function isMatchMetaCompatible(existing = {}, requested = {}, gameType = '') {
+  const normalizedGameType = String(gameType || '').trim().toLowerCase();
+
+  if (normalizedGameType === 'poolroyale') {
+    const existingVariant = existing?.variant || '';
+    const requestedVariant = requested?.variant || '';
+    // Pool Royale quick matchmaking now only partitions by stake + variant.
+    // Stake is checked outside this helper (see getAvailableTable).
+    // Keep missing variants as wildcards for backward compatibility.
+    if (!existingVariant || !requestedVariant) return true;
+    return existingVariant === requestedVariant;
+  }
+
+  const allKeys = new Set([
+    ...Object.keys(existing || {}),
+    ...Object.keys(requested || {})
+  ]);
+  for (const key of allKeys) {
+    if (key === 'preferredSide') {
+      // Side is a seat preference, not a matchmaking partition. Keep players
+      // in the same queue even when they pick opposite colors.
+      continue;
+    }
+    const existingValue = existing?.[key];
+    const requestedValue = requested?.[key];
+    // Treat missing keys as wildcards so players can still pair when one client
+    // sends less detailed lobby metadata (for example missing tableSize).
+    if (!existingValue || !requestedValue) continue;
+    if (existingValue !== requestedValue) return false;
+  }
+  return true;
+}
+
 function isRateLimited(socket, key, cooldownMs) {
   const now = Date.now();
   const last = lastActionBySocket.get(socket.id)?.[key] || 0;
@@ -558,8 +638,7 @@ function createLobbyTable({
   gameType,
   stake = 0,
   maxPlayers = 4,
-  meta = {},
-  matchKey = ''
+  meta = {}
 }) {
   const key = `${gameType}-${maxPlayers}`;
   if (!lobbyTables[key]) lobbyTables[key] = [];
@@ -571,8 +650,7 @@ function createLobbyTable({
     players: [],
     currentTurn: null,
     ready: new Set(),
-    meta,
-    matchKey
+    meta
   };
   lobbyTables[key].push(table);
   tableMap.set(table.id, table);
@@ -590,13 +668,6 @@ function getAvailableTable(
   forcedTableId = null
 ) {
   const normalizedMeta = normalizeMatchMeta(matchMeta);
-  const requestedMatchKey = buildLobbyMatchKey({
-    gameType,
-    stake,
-    maxPlayers,
-    matchMeta: normalizedMeta,
-    forcedTableId: forcedTableId || ''
-  });
   const key = `${gameType}-${maxPlayers}`;
   if (!lobbyTables[key]) lobbyTables[key] = [];
   if (forcedTableId) {
@@ -619,27 +690,20 @@ function getAvailableTable(
         gameType,
         stake,
         maxPlayers,
-        meta: normalizedMeta,
-        matchKey: requestedMatchKey
+        meta: normalizedMeta
       });
     }
     // Ignore stale/non-existent forced ids so quick matchmaking can still pair
     // users by game type + stake instead of trapping them in a private table.
   }
-  const open = lobbyTables[key].find((t) => {
-    if (isReservedHostedTableId(t.id, gameType, maxPlayers)) return false;
-    if (t.stake !== stake || t.players.length >= t.maxPlayers) return false;
-    if (t.matchKey) return t.matchKey === requestedMatchKey;
-    return isMatchMetaCompatible(t.meta, normalizedMeta, gameType);
-  });
+  const open = lobbyTables[key].find(
+    (t) =>
+      t.stake === stake &&
+      t.players.length < t.maxPlayers &&
+      isMatchMetaCompatible(t.meta, normalizedMeta, gameType)
+  );
   if (open) return open;
-  return createLobbyTable({
-    gameType,
-    stake,
-    maxPlayers,
-    meta: normalizedMeta,
-    matchKey: requestedMatchKey
-  });
+  return createLobbyTable({ gameType, stake, maxPlayers, meta: normalizedMeta });
 }
 
 function resolveSeatIdentityFromTableId(tableId, fallbackGameType, fallbackMaxPlayers) {
